+++ src/main/java/org/starcoin/search/repository/TokenVolumeDTO.java
@@ -1,8 +1,10 @@
 package org.starcoin.search.repository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface TokenVolumeDTO {
     public BigDecimal getVolume();
-    public BigDecimal getVolumeAmount();
+
+    public BigInteger getVolumeAmount();
 }
